==> 01_DataLoading_Rob.ipynb <==
import pandas as pd
import numpy as np

## Fetch data from internet

!rm -rf ./data
!mkdir data

print('Getting movielens data...\n')
!wget http://files.grouplens.org/datasets/movielens/ml-20m.zip -O ./data/ml-20m.zip

print('Getting imdb title data...\n')
!wget https://datasets.imdbws.com/title.akas.tsv.gz -O ./data/title.akas.tsv.gz

print('Getting imdb ratings data...\n')
!wget https://datasets.imdbws.com/title.ratings.tsv.gz -O ./data/title.ratings.tsv.gz

print('Getting imdb title data...\n')
!wget https://datasets.imdbws.com/title.basics.tsv.gz -O ./data/title.basics.tsv.gz
print('Data downloads complete.\n')

print('Unzipping...')
!unzip -o ./data/ml-20m.zip -d ./data
#!mkdir ./data/title.akas/
#!gunzip -c ./data/title.akas.tsv.gz > ./data/title.akas/title.akas.tsv
!mkdir ./data/title.ratings/
!gunzip -c ./data/title.ratings.tsv.gz > ./data/title.ratings/title.ratings.tsv
!mkdir ./data/title.basics/
!gunzip -c ./data/title.basics.tsv.gz > ./data/title.basics/title.basics.tsv
print('Decompression complete')

## Load data into Pandas dataframe

df_movies = pd.read_csv('./data/ml-20m/movies.csv')
df_rating = pd.read_csv('./data/ml-20m/ratings.csv')
df_tags = pd.read_csv('./data/ml-20m/tags.csv')
df_links = pd.read_csv('./data/ml-20m/links.csv')
df_genome_scores = pd.read_csv('./data/ml-20m/genome-scores.csv')
df_genome_tags = pd.read_csv('./data/ml-20m/genome-tags.csv')

imdb_title = pd.read_table('./data/title.akas/title.akas.tsv', dtype={'titleId':str,
                                                                      'ordering':int,
                                                                      'title':str,
                                                                      'region':str,
                                                                      'language':str,
                                                                      'types':str,
                                                                      'attributes':str,
                                                                      'isOriginalTitle':str}, na_values = '\\N')
imdb_ratings = pd.read_table('./data/title.ratings/title.ratings.tsv', na_values = '\\N')
imdb_title_basics = pd.read_table('./data/title.basics/title.basics.tsv', na_values = '\\N', dtype={'tconst':str, 
                                                                                                    'titleType':str, 
                                                                                                    'primaryTitle':str, 
                                                                                                    'originalTitle':str, 
                                                                                                    'isAdult':str, 
                                                                                                    'startYear':str, 
                                                                                                    'endYear':str, 
                                                                                                    'runtimeMinutes':str, 
                                                                                                    'genres':str})

## Some preprocessing steps from here

df_rating.timestamp = pd.to_datetime(df_rating.timestamp,unit='s')

df_tags.timestamp = pd.to_datetime(df_tags.timestamp,unit='s')

df_movies['movie_year'] = df_movies.title.str.extract('\(([0-9][0-9][0-9][0-9])\)')
df_movies = pd.concat([df_movies, df_movies.genres.str.get_dummies(sep='|')  ], axis=1)

#### Format imdb_ratings for joining

imdb_ratings['imdbId'] = pd.to_numeric(imdb_ratings.tconst.str.replace('tt(0)*', ''))
imdb_ratings = imdb_ratings.drop(['tconst'], axis=1)

display(imdb_ratings.head())

#### Format imdb_title_basics for joining

imdb_title_basics['imdbId'] = pd.to_numeric(imdb_title_basics.tconst.str.replace('tt(0)*', ''))
imdb_title_basics = imdb_title_basics.drop(['tconst','endYear'], axis=1)
imdb_title_basics['genre_list'] = imdb_title_basics['genres'].str.split(',')
display(imdb_title_basics.head())

### Merge IMDB files
#### imdb_title_basics with df_links

film_database_data = pd.merge(df_links, imdb_title_basics, how='left', on=['imdbId'])

print("Merged dataframe shape: {}".format(film_database_data.shape))
print("Original links shape: {}".format(df_links.shape))

print("\nMerged dataframe:")
film_database_data.head()

#### film_database_data with imdb_ratings

df_imdb_film_data_with_ratings = pd.merge(film_database_data, imdb_ratings, how='left', on=['imdbId'])

df_imdb_film_data_with_ratings.head()

## Save to pickle file

df_movies.to_pickle("./data/movies.pkl")

df_rating.to_pickle("./data/rating.pkl")

df_tags.to_pickle("./data/tags.pkl")

df_links.to_pickle("./data/links.pkl")

df_genome_scores.to_pickle("./data/genome-scores.pkl")

df_genome_tags.to_pickle("./data/genome-tags.pkl")

df_imdb_film_data_with_ratings.to_pickle("./data/imdb-data.pkl")
