==> 03_EmbeddingExperiment.ipynb <==
## Embedding experiment

import pandas as pd
import tensorflow as tf
import numpy as np
import shutil
from sklearn.model_selection import train_test_split
from google.datalab.ml import TensorBoard
print(tf.__version__)

tf.logging.set_verbosity(tf.logging.INFO)

### Variables

OUTDIR = 'embedding_experiment'
shutil.rmtree(OUTDIR, ignore_errors = True) # start fresh each time

### Reading data using Pandas

df_movies = pd.read_pickle("./data/movies.pkl")
df_rating = pd.read_pickle("./data/rating.pkl")
df_tags = pd.read_pickle("./data/tags.pkl")
df_links = pd.read_pickle("./data/links.pkl")
df_genome_scores = pd.read_pickle("./data/genome-scores.pkl")
df_genome_tags = pd.read_pickle("./data/genome-tags.pkl")

### Basic test & training set construction

df_dataset = df_rating.join(df_movies, on=['movieId'], lsuffix='_rating', rsuffix='_movies',how='inner')
df_dataset = df_dataset.reset_index(drop=True)
df_dataset.columns

msk = np.random.rand(len(df_dataset)) < 0.8
df_train = df_dataset[msk]
df_test = df_dataset[~msk]

print("Test length: %s , Training length: %s" % (len(df_test),len(df_train)))

FEATURES = list(df_dataset.columns)
FEATURES.remove('movieId_rating')
FEATURES.remove('timestamp')
FEATURES.remove('movieId_movies')
FEATURES.remove('(no genres listed)')
FEATURES.remove('title') ## for now, no textual features without tokenizers
FEATURES.remove('genres')
LABEL = FEATURES.pop(1)

print("Features: %s \r\n" % (FEATURES))
print("Label: %s \r\n" % (LABEL))
print("Feature Col types: %s \r\n" % df_train[FEATURES].dtypes)
print("Label Col types: %s \r\n" % df_train[LABEL].dtypes)

### Input functions

def make_train_input_fn(df, num_epochs):
  return tf.estimator.inputs.pandas_input_fn(
    x = df[FEATURES],
    y = df[LABEL],
    batch_size = 128,
    num_epochs = num_epochs,
    shuffle = True,
    queue_capacity = 1000
  )


def make_eval_input_fn(df):
  return tf.estimator.inputs.pandas_input_fn(
    x = df[FEATURES],
    y = df[LABEL],
    batch_size = 128,
    shuffle = False,
    queue_capacity = 1000
  )

def serving_input_fn(df):
  return tf.estimator.inputs.pandas_input_fn(
    x = df[FEATURES],
    y = None,
    batch_size = 128,
    shuffle = False,
    queue_capacity = 1000
  )

### Create feature columns

def make_feature_cols():
  input_columns = []
  ## input_columns = [tf.feature_column.numeric_column(k) for k in FEATURES] -- No fields are fully numerical
  
  ## The binary fields:
  binary_fields = ['Action', 'Adventure', 'Animation', 'Children', 'Comedy', 'Crime', 'Documentary', 'Drama', 'Fantasy', 'Film-Noir', 'Horror', 
                   'IMAX', 'Musical', 'Mystery', 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western'] 
  
  binary_columns = [tf.feature_column.categorical_column_with_identity(k, num_buckets=2,default_value=0) for k in binary_fields]

  ## Embedding the category fields
  #input_columns = tf.feature_column.embedding_column(categorical_column=binary_columns,dimension=10)
  
  input_columns = [tf.feature_column.embedding_column(tf.feature_column.crossed_column(binary_columns,20**2),dimension=5)]
  
  return input_columns #binary_columns #input_columns

print(make_feature_cols())

### Define model train & evaluate function

def train_and_evaluate(output_dir, num_train_steps):
  
    EVAL_INTERVAL = 300
    run_config = tf.estimator.RunConfig(save_checkpoints_secs = EVAL_INTERVAL,
                                      keep_checkpoint_max = 3)
    #estimator = tf.estimator.LinearRegressor(
    #                   model_dir = output_dir,
    #                   feature_columns = make_feature_cols())
    
    estimator = tf.estimator.DNNRegressor(
                       model_dir = output_dir,
                       feature_columns = make_feature_cols(),
                       hidden_units = [64, 32,16,8],
                       config = run_config)
    
    train_spec=tf.estimator.TrainSpec(
                       input_fn = make_train_input_fn(df_train, num_epochs = 10),
                       max_steps = num_train_steps)

    #exporter = tf.estimator.LatestExporter('exporter', serving_input_fn)

    eval_spec=tf.estimator.EvalSpec(
                       input_fn = make_eval_input_fn(df_test),
                       steps = None,
                       start_delay_secs = 1, # start evaluating after N seconds
                       throttle_secs = EVAL_INTERVAL,  # evaluate every N seconds
                       #exporters = exporter
    )
    
    tf.estimator.train_and_evaluate(estimator, train_spec, eval_spec)
    
    return(estimator)

### Run!

# Run training    
shutil.rmtree(OUTDIR, ignore_errors = True) # start fresh each time
model = train_and_evaluate(OUTDIR, num_train_steps = 1000)

def print_rmse(model, df):
  metrics = model.evaluate(input_fn = make_eval_input_fn(df))
  print('RMSE on dataset = {}'.format(np.sqrt(metrics['average_loss'])))
#print_rmse(model, df_valid) ## Normally we want to valide on a separate validation set
print_rmse(model, df_test)

### TensorflowBoard for debugging purposes

TensorBoard().start(OUTDIR)

### Shutting down TensorflowBoard

# to list Tensorboard instances
TensorBoard().list()

pids_df = TensorBoard.list()
if not pids_df.empty:
    for pid in pids_df['pid']:
        TensorBoard().stop(pid)
        print('Stopped TensorBoard with pid {}'.format(pid))
