==> 04_WALS_MatrixFactorization.ipynb <==
## Wals MatrixFactorization for embedding

import pandas as pd
import tensorflow as tf

from tensorflow.python.lib.io import file_io
from tensorflow.contrib.factorization import WALSMatrixFactorization

import numpy as np
import shutil
from sklearn.model_selection import train_test_split
from scipy.sparse import coo_matrix
from google.datalab.ml import TensorBoard
print(tf.__version__)

tf.logging.set_verbosity(tf.logging.INFO)

### Setting variables

TEST_SET_RATIO = 0.1
OUTDIR = 'wals_experiment'
shutil.rmtree(OUTDIR, ignore_errors = True) # start fresh each time

DEFAULT_PARAMS = {
    'weights': True,
    'latent_factors': 5,
    'num_iters': 20,
    'regularization': 0.07,
    'unobs_weight': 0.01,
    'wt_type': 0,
    'feature_wt_factor': 130.0,
    'feature_wt_exp': 0.08,
    'delimiter': '\t'
}

OPTIMIZED_PARAMS = {
    'latent_factors': 34,
    'regularization': 9.83,
    'unobs_weight': 0.001,
    'feature_wt_factor': 189.8,
}

for key in OPTIMIZED_PARAMS:
  DEFAULT_PARAMS[key] = OPTIMIZED_PARAMS[key]

### Reading data with Pandas

df_movies = pd.read_pickle("./data/movies.pkl")
df_rating = pd.read_pickle("./data/rating.pkl")
df_tags = pd.read_pickle("./data/tags.pkl")
df_links = pd.read_pickle("./data/links.pkl")
df_genome_scores = pd.read_pickle("./data/genome-scores.pkl")
df_genome_tags = pd.read_pickle("./data/genome-tags.pkl")

df_rating.describe(include = 'all')

df_rating['rating'] = pd.to_numeric(df_rating['rating'], errors='coerce')
df_rating['userId'] = pd.to_numeric(df_rating['userId'], errors='coerce')
df_rating['movieId'] = pd.to_numeric(df_rating['movieId'], errors='coerce')
df_rating = df_rating.dropna(subset=['rating','userId','movieId'])
df_rating['rating'] = df_rating['rating'].astype(float)
df_rating['userId'] = df_rating['userId'].astype(int)
df_rating['movieId'] = df_rating['movieId'].astype(int)
del df_rating['timestamp']
df_rating.describe(include = 'all')

### Prepare dataset

ratings = df_rating.as_matrix(['userId', 'movieId', 'rating'])

ratings[:10]

U_ratings = df_rating.groupby(['userId'])['rating'].apply(np.array)

U_movies= df_rating.groupby(['userId'])['movieId'].apply(np.array)

U = pd.concat([U_ratings, U_movies], axis=1)

del U_ratings
del U_movies

U['rating'][1]

a1 = np.array(U.rating[1])
a2 = np.array(U.movieId[1])
print(a1)
print(a2)
print()
a3 = take_topN(a1, a2, 3)
print(a3)

print(take_topN( np.array([3.5, 4.5, 2.5, 6.7]), np.array([100, 50, 150, 2500]), 3))

def take_topN(array1, array2, n=3):
  indices = np.argsort(array1)#[::-1]
  return array2[indices[0:n]]

U['col2'] = U.apply(lambda row: list(take_topN(row['rating'],row['movieId'],3)),axis=1)

U['col2'] = U.apply(lambda row: list(take_topN(row['rating'],row['movieId'],3)),axis=1)

U.head(20)

#U['topN'] = U.apply(lambda row: take_topN(row['rating'], row['movieId']))

#U_total = pd.merge(U_ratings.to_frame(), U_movies.to_frame(), on='userId', how='inner')

del U_ra

np_items = df_rating.movieId.as_matrix()
np_users = df_rating.userId.as_matrix()
unique_users = np.unique(np_users)
unique_items = np.unique(np_items)
n_users = unique_users.shape[0]
n_items = unique_items.shape[0]
max_item = unique_items[-1]
max_user = unique_users[-1]

if n_users != max_user or n_items != max_item:
  z = np.zeros(max_user + 1, dtype=int)
  z[unique_users] = np.arange(n_users)
  u_r = z[np_users]
  z = np.zeros(max_item + 1, dtype=int)
  z[unique_items] = np.arange(n_items)
  i_r = z[np_items]
  np_ratings = df_rating.rating.as_matrix()
  ratings = np.zeros((np_ratings.shape[0], 3), dtype=object)
  ratings[:, 0] = u_r
  ratings[:, 1] = i_r
  ratings[:, 2] = np_ratings
else:
    ratings = df_rating.as_matrix(['user_id', 'item_id', 'rating'])
    ratings[:, 0] -= 1
    ratings[:, 1] -= 1

print('Number of items: {} and max item {} / Number of users: {}').format(n_items,max_item,n_users)

test_set_size = int(len(ratings) * TEST_SET_RATIO)
test_set_idx = np.random.choice(xrange(len(ratings)),
                                size=test_set_size, replace=False)
test_set_idx = sorted(test_set_idx)

ts_ratings = ratings[test_set_idx]
tr_ratings = np.delete(ratings, test_set_idx, axis=0)

u_tr, i_tr, r_tr = zip(*tr_ratings)
u_ts, i_ts, r_ts = zip(*ts_ratings)

print(len(r_tr))
print(len(u_tr))
print(len(i_tr))
print(n_users * n_items)

tr_sparse = coo_matrix((r_tr, (u_tr, i_tr)), shape=(n_users, n_items))
test_sparse = coo_matrix((r_ts, (u_ts, i_ts)), shape=(n_users, n_items))

user_sparse_tensor = tf.SparseTensor(indices=tr_sparse.row, dense_shape=(tr_sparse.shape[0],1))

input_tensor = tf.SparseTensor(indices=zip(tr_sparse.row, tr_sparse.col),
                                values=(tr_sparse.data).astype(np.float32),
                                dense_shape=tr_sparse.shape)

input_tensor.

print(input_tensor.get_shape())

limit = input_tensor.indices
print(limit)

def train_and_evaluate(args):
    train_steps = int(0.5 + (1.0 * args['num_epochs'] * args['nusers']) / args['batch_size'])
    steps_in_epoch = int(0.5 + args['nusers'] / args['batch_size'])
    print('Will train for {} steps, evaluating once every {} steps'.format(train_steps, steps_in_epoch))
    def experiment_fn(output_dir):
        return tf.contrib.learn.Experiment(
            tf.contrib.factorization.WALSMatrixFactorization(
                         num_rows=args['nusers'], num_cols=args['nitems'],
                         embedding_dimension=args['n_embeds'],
                         model_dir=args['output_dir']),
            train_input_fn=read_dataset(tf.estimator.ModeKeys.TRAIN, args),
            eval_input_fn=read_dataset(tf.estimator.ModeKeys.EVAL, args),
            train_steps=train_steps,
            eval_steps=1,
            min_eval_frequency=steps_in_epoch
        )

    from tensorflow.contrib.learn.python.learn import learn_runner
    learn_runner.run(experiment_fn, args['output_dir'])
    
    batch_predict(args)

import shutil
shutil.rmtree(OUTDIR, ignore_errors=True)
train_and_evaluate({
    'output_dir': OUTDIR,
    #'input_path': 'gs://{}/wals/preproc_tft'.format(BUCKET),
    'num_epochs': 0.05,
    'nitems': 5668,
    'nusers': 82802,

    'batch_size': 512,
    'n_embeds': 10,
    'topk': 3
  })
