==> 99_TMD_Data Collection.ipynb <==
!pip install tmdbsimple

import requests
import pandas as pd
import tmdbsimple as tmdb
import pickle

api_key='9d4724b28cde31159c7465392dbcf25c'
access_token='eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI5ZDQ3MjRiMjhjZGUzMTE1OWM3NDY1Mzky\
ZGJjZjI1YyIsInN1YiI6IjViZWMwYjFjOTI1MTQxMDI2NzAwN2Y1ZSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIj\
oxfQ.lMnhij5psVXIf8XfN5VFkrcfU94oWl3KBMj2ggm-VBM'
review_and_keywords = {}

tmdb.API_KEY = api_key

df_links = pickle.load(open("./data/links.pkl", 'rb'))
df_links['tmdbId']
df_links.head()

# Function to create record from movie id

Stats about movie

Keywords

Longest review

movie = tmdb.Movies(603)

## Stats

response_keys_for_df = ['title', 'budget', 'revenue', 'genres', 'release_date', 'popularity', 'vote_average', 'vote_count', 'overview', 'poster_path']
movie_dict = {x:movie.info()[x] for x in response_keys_for_df}
movie_dict['genres'] = [genre['name'] for genre in movie_dict['genres']]

## Keywords

keywords = []
for word in movie.keywords()['keywords']:
  keywords.append(word['name'])
print(keywords)

## Longest Review

#for review in movie.reviews()['results']
#movie.reviews()['results']
reviews = [(review['content']) for review in movie.reviews()['results']]
longest_review = max(reviews, key=len)

# Add keywords and review to dict

review_and_keywords[movie.id] = {'longest_review':longest_review, 'keywords':keywords}

pd.DataFrame.from_dict(movie_dict)

review_and_keywords
